+++ fidget/src/core/vm/mod.rs
@@ -4,10 +4,10 @@ use crate::{
     context::Node,
     eval::{
         BulkEvaluator, Function, MathFunction, Tape, Trace, TracingEvaluator,
-        VarMap,
     },
     shape::{RenderHints, Shape},
     types::{Grad, Interval},
+    var::VarMap,
     Context, Error,
 };
 use std::sync::Arc;
@@ -177,6 +177,10 @@ impl<const N: usize> Function for GenericVmFunction<N> {
     fn size(&self) -> usize {
         GenericVmFunction::size(self)
     }
+
+    fn vars(&self) -> &VarMap<usize> {
+        &self.0.vars
+    }
 }
 
 impl<const N: usize> RenderHints for GenericVmFunction<N> {
@@ -190,9 +194,9 @@ impl<const N: usize> RenderHints for GenericVmFunction<N> {
 }
 
 impl<const N: usize> MathFunction for GenericVmFunction<N> {
-    fn new(ctx: &Context, node: Node) -> Result<(Self, VarMap), Error> {
-        let (d, vs) = VmData::new(ctx, node)?;
-        Ok((Self(Arc::new(d)), vs))
+    fn new(ctx: &Context, node: Node) -> Result<Self, Error> {
+        let d = VmData::new(ctx, node)?;
+        Ok(Self(d.into()))
     }
 }