+++ fidget/src/jit/x86_64/interval.rs
@@ -85,8 +85,8 @@ impl Assembler for IntervalAssembler {
             ; movq [rsp + sp_offset], Rx(reg(src_reg))
         );
     }
-    fn build_input(&mut self, out_reg: u8, src_arg: u8) {
-        let pos = 8 * (src_arg as i32);
+    fn build_input(&mut self, out_reg: u8, src_arg: u32) {
+        let pos = 8 * i32::try_from(src_arg).unwrap();
         dynasm!(self.0.ops
             ; vmovq Rx(reg(out_reg)), [rdi + pos]
         );